feat: delete saga ci: separated testing for the src, added utils for testing clarity feat: actions, reducers, and sagas no longer carry the "rest" prefix in the file names
