+++ babel.config.js
@@ -7,8 +7,7 @@ module.exports = function (api) {
         {
           targets: {
             node: 'current'
-          },
-          useBuiltIns: 'usage'
+          }
         }
       ]
     ]
